Updated for MySQL; removed seminar classes from recommendations; cleaned up static files

+++ recommend/models.py
@@ -7,7 +7,7 @@ DEFAULT_RECOMMENDATION_TYPE = "for-you"
 
 # Create your models here.
 class Rating(models.Model):
-    user_id = models.IntegerField(default=0)
+    user_id = models.BigIntegerField(default=0)
     subject_id = models.CharField(max_length=50)
     value = models.IntegerField(default=0)
 
@@ -15,7 +15,7 @@ class Rating(models.Model):
         return "User {} rated {} as {}".format(self.user_id, self.subject_id, self.value)
 
 class Recommendation(models.Model):
-    user_id = models.IntegerField(default=0)
+    user_id = models.BigIntegerField(default=0)
     rec_type = models.CharField(max_length=20)
     subjects = models.CharField(max_length=500)
